Enable post creation and integrate Firestore backend
Added navigation and UI for creating posts, connected CreatePostScreen to Firestore for saving posts, and updated validation to limit posts to 250 characters. FeedScreen now shows an empty state when there are no posts. Firebase authentication is now initialized with React Native persistence for better session handling.

Dealing with a firebase storage issue. Have to re-config the DB rules and permissions.